+++ AddOns/Auctionator/Source_Mainline/Config/Mixins/Cancelling.lua
@@ -4,7 +4,7 @@ function AuctionatorConfigCancellingFrameMixin:OnLoad()
   Auctionator.Debug.Message("AuctionatorConfigCancellingFrameMixin:OnLoad()")
 
   self.name = AUCTIONATOR_L_CONFIG_CANCELLING_CATEGORY
-  self.parent = AUCTIONATOR_L_ADDONS_AUCTIONATOR
+  self.parent = "Auctionator"
 
   self:SetupPanel()
 end
@@ -12,6 +12,7 @@ end
 function AuctionatorConfigCancellingFrameMixin:OnShow()
   self.UndercutScanPetsGear:SetChecked(Auctionator.Config.Get(Auctionator.Config.Options.UNDERCUT_SCAN_NOT_LIFO))
   self.UndercutScanMatchIlvlVariants:SetChecked(Auctionator.Config.Get(Auctionator.Config.Options.UNDERCUT_SCAN_GEAR_MATCH_ILVL_VARIANTS))
+  self.UndercutItemsAhead:SetNumber(Auctionator.Config.Get(Auctionator.Config.Options.UNDERCUT_ITEMS_AHEAD))
   self.CancelUndercutShortcut:SetShortcut(Auctionator.Config.Get(Auctionator.Config.Options.CANCEL_UNDERCUT_SHORTCUT))
 end
 
@@ -20,6 +21,7 @@ function AuctionatorConfigCancellingFrameMixin:Save()
 
   Auctionator.Config.Set(Auctionator.Config.Options.UNDERCUT_SCAN_NOT_LIFO, self.UndercutScanPetsGear:GetChecked())
   Auctionator.Config.Set(Auctionator.Config.Options.UNDERCUT_SCAN_GEAR_MATCH_ILVL_VARIANTS, self.UndercutScanMatchIlvlVariants:GetChecked())
+  Auctionator.Config.Set(Auctionator.Config.Options.UNDERCUT_ITEMS_AHEAD, self.UndercutItemsAhead:GetNumber())
   Auctionator.Config.Set(Auctionator.Config.Options.CANCEL_UNDERCUT_SHORTCUT, self.CancelUndercutShortcut:GetShortcut())
 end