+++ AddOns/Auctionator/Source_Mainline/Config/Mixins/Quantities.lua
@@ -4,7 +4,7 @@ function AuctionatorConfigQuantitiesFrameMixin:OnLoad()
   Auctionator.Debug.Message("AuctionatorConfigQuantitiesFrameMixin:OnLoad()")
 
   self.name = AUCTIONATOR_L_CONFIG_QUANTITIES_CATEGORY
-  self.parent = AUCTIONATOR_L_ADDONS_AUCTIONATOR
+  self.parent = "Auctionator"
 
   self:SetupPanel()