+++ AddOns/Auctionator/Source_Mainline/Config/Mixins/Selling.lua
@@ -4,7 +4,7 @@ function AuctionatorConfigSellingFrameMixin:OnLoad()
   Auctionator.Debug.Message("AuctionatorConfigSellingFrameMixin:OnLoad()")
 
   self.name = AUCTIONATOR_L_CONFIG_SELLING_CATEGORY
-  self.parent = AUCTIONATOR_L_ADDONS_AUCTIONATOR
+  self.parent = "Auctionator"
 
   self:SetupPanel()
 end