+++ AddOns/Auctionator/Source_Mainline/Config/Mixins/SellingAllItems.lua
@@ -4,7 +4,7 @@ function AuctionatorConfigSellingAllItemsFrameMixin:OnLoad()
   Auctionator.Debug.Message("AuctionatorConfigSellingAllItemsFrameMixin:OnLoad()")
 
   self.name = AUCTIONATOR_L_CONFIG_SELLING_ALL_ITEMS_CATEGORY
-  self.parent = AUCTIONATOR_L_ADDONS_AUCTIONATOR
+  self.parent = "Auctionator"
 
   self:SetupPanel()