+++ AddOns/Auctionator/Source_Mainline/Config/Mixins/SellingShortcuts.lua
@@ -4,7 +4,7 @@ function AuctionatorConfigSellingShortcutsFrameMixin:OnLoad()
   Auctionator.Debug.Message("AuctionatorConfigSellingShortcutsFrameMixin:OnLoad()")
 
   self.name = AUCTIONATOR_L_CONFIG_SELLING_SHORTCUTS_CATEGORY
-  self.parent = AUCTIONATOR_L_ADDONS_AUCTIONATOR
+  self.parent = "Auctionator"
 
   self:SetupPanel()
 end
@@ -16,6 +16,7 @@ function AuctionatorConfigSellingShortcutsFrameMixin:OnShow()
 
   self.PostShortcut:SetShortcut(Auctionator.Config.Get(Auctionator.Config.Options.SELLING_POST_SHORTCUT))
   self.SkipShortcut:SetShortcut(Auctionator.Config.Get(Auctionator.Config.Options.SELLING_SKIP_SHORTCUT))
+  self.PrevShortcut:SetShortcut(Auctionator.Config.Get(Auctionator.Config.Options.SELLING_PREV_SHORTCUT))
 end
 
 function AuctionatorConfigSellingShortcutsFrameMixin:Save()
@@ -27,6 +28,7 @@ function AuctionatorConfigSellingShortcutsFrameMixin:Save()
 
   Auctionator.Config.Set(Auctionator.Config.Options.SELLING_POST_SHORTCUT, self.PostShortcut:GetShortcut())
   Auctionator.Config.Set(Auctionator.Config.Options.SELLING_SKIP_SHORTCUT, self.SkipShortcut:GetShortcut())
+  Auctionator.Config.Set(Auctionator.Config.Options.SELLING_PREV_SHORTCUT, self.PrevShortcut:GetShortcut())
 end
 
 function AuctionatorConfigSellingShortcutsFrameMixin:UnhideAllClicked()