+++ AddOns/Auctionator/Source_Mainline/Tabs/Selling/Mixins/BagItemSelected.lua
@@ -1,5 +1,12 @@
 AuctionatorBagItemSelectedMixin = CreateFromMixins(AuctionatorBagItemMixin)
 
+function AuctionatorBagItemSelectedMixin:SetItemInfo(info, ...)
+  AuctionatorBagItemMixin.SetItemInfo(self, info, ...)
+  self.IconSelectedHighlight:Hide()
+  self.IconBorder:SetShown(info ~= nil)
+  self.Icon:SetAlpha(1)
+end
+
 function AuctionatorBagItemSelectedMixin:OnClick(button)
   local wasCursorItem = C_Cursor.GetCursorItem()
   if not self:ProcessCursor() then